// src/Types/GeoPoint.ts
export declare interface GeoPoint {
  /**
   * Measurement of a location's distance east or west of the Earth's Prime Meridian, which is located at 0 degrees. Longitude values range from -180 degrees on one side of the Earth to +180 degrees on the other side.
   */
  lon: number;

  /**
   * Measurement of a location's distance north or south of the Earth's equator, which is located at 0 degrees. Latitude values range from -90 degrees at the South Pole to +90 degrees at the North Pole.
   */
  lat: number;
}